refactor(dh): move download to service
