+++ libs/dh/market-participant/actors/feature-actors/src/lib/drawer/actor-audit-log-tab/dh-actor-audit-log-tab.component.ts
@@ -85,6 +85,7 @@ export class DhActorAuditLogTabComponent {
       .map((x) => x.trim());
 
     if (!values) return {};
+
     const [payloadGln, payloadStartsAt, payloadGridArea, payloadProcessType, payloadStopsAt] =
       values;
     const actorNumberName = this.actorNumberNameLookup()[payloadGln];
