+++ libs/dh/market-participant/actors/feature-actors/src/lib/drawer/balance-responsible-relation-tab/dh-balance-responsible-relation-tab.component.html
@@ -29,7 +29,7 @@ limitations under the License.
       (filtersChanges)="store.updateFilters($event)"
     />
 
-    <watt-button icon="download" variant="text" (click)="download()">{{
+    <watt-button icon="download" variant="text" (click)="store.download()">{{
       "shared.download" | transloco
     }}</watt-button>
   </vater-stack>
